add support for write_version_directive, which cascades some required changes throughout the Emitter, Scanner and Parser.
Note that in Parser I found an incomplete process_directives().
I stubbed in the rest of what is needed but left tag_handles
unhandled, as I don't really know what those are (yet).

Also, while python is using tuples for the yaml_version, I've
been using a string. This seems to be working out okay. Changing
to a two element list should be okay, but seemed out of the idiom.